# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lotty A/B & feature flags API
  description: |
    HTTP API платформы для А/Б тестирования.

    **Идентификация (демо):** заголовки `X-User-ID` и `X-Role`.
    Если `X-Role` не передан, роль берётся из записи пользователя в БД по `X-User-ID`
    (по умолчанию `X-User-ID` пустой → `"viewer"`).
    Роли: `admin`, `experimenter`, `approver`, `viewer`.

    Ошибки: JSON `{ "error": "<текст>" }` со статусом 400/401/403/404/405/500 и т.д.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер (порт задаётся `APP_PORT`, по умолчанию 8080)
...
paths:
...
  /v1/flags:
    get:
      tags: [flags]
      summary: Список флагов
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flag"
        "500":
          $ref: "#/components/responses/Error"
    post:
      tags: [flags]
      summary: Создать флаг (idempotent при `Idempotency-Key`)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagCreate"
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flag"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "405":
          description: Method not allowed
...
  /v1/experiments:
    get:
      tags: [experiments]
      summary: Список экспериментов
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experiment"
    post:
      tags: [experiments]
      summary: Создать эксперимент (idempotent при `Idempotency-Key`)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentCreate"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/decide:
    post:
      tags: [decide]
      summary: Получить варианты для субъекта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecideRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [decisions]
                properties:
                  decisions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Decision"
        "400":
          $ref: "#/components/responses/Error"
        "405":
          description: Method not allowed
  /v1/events/batch:
    post:
      tags: [events]
      summary: Пакетная доставка событий
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  items:
                    $ref: "#/components/schemas/EventIngest"
      responses:
        "200":
          description: Итоги обработки батча
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBatchResult"
        "400":
          $ref: "#/components/responses/Error"
...
  /v1/admin/users:
    post:
      tags: [admin]
      summary: Создать/обновить пользователя (только admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, role]
              properties:
                id:
                  type: string
                role:
                  $ref: "#/components/schemas/Role"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/approver-policy:
    post:
      tags: [admin]
      summary: Политика согласующих (только admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproverPolicy"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproverPolicy"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/event-types:
    get:
      tags: [admin]
      summary: Список типов событий
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventType"
    post:
      tags: [admin]
      summary: Создать тип события (только admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventType"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventType"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/replay:
    post:
      tags: [admin]
      summary: Пересчёт атрибуции за интервал (только admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [experiment_id, from, to]
              properties:
                experiment_id:
                  type: string
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  rebuilt:
                    type: integer
                    description: Число пересобранных attributed-событий (смысл поля в коде сервиса)
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Role:
      type: string
      enum: [admin, experimenter, approver, viewer]
    User:
      type: object
      properties:
        id:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    FlagCreate:
      type: object
      required: [key, type, default_value]
      properties:
        key:
          type: string
        type:
          type: string
          enum: [string, number, bool]
        default_value:
          description: JSON literal
        description:
          type: string
    Flag:
      allOf:
        - $ref: "#/components/schemas/FlagCreate"
        - type: object
          properties:
            owner:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    Variant:
      type: object
      required: [name, value, weight, is_control]
      properties:
        name:
          type: string
        value:
          description: JSON literal
        weight:
          type: integer
        is_control:
          type: boolean
    GuardrailRule:
      type: object
      required: [metric_key, threshold, window_seconds, action]
      properties:
        metric_key:
          type: string
        threshold:
          type: number
        window_seconds:
          type: integer
        action:
          type: string
          enum: [pause]
    RuleNode:
      type: object
      properties:
        op:
          type: string
        field:
          type: string
        value: {}
        values:
          type: array
          items: {}
        children:
          type: array
          items:
            $ref: "#/components/schemas/RuleNode"
    ExperimentCreate:
      type: object
      required: [flag_key, name, audience_percent, variants]
      properties:
        flag_key:
          type: string
        name:
          type: string
        audience_percent:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        targeting:
          $ref: "#/components/schemas/RuleNode"
        guardrails:
          type: array
          items:
            $ref: "#/components/schemas/GuardrailRule"
...
    Experiment:
      type: object
      properties:
        id:
          type: string
        flag_key:
          type: string
        name:
          type: string
        owner_id:
          type: string
        status:
          type: string
          description: draft, in_review, approved, running, paused, completed, archived, rejected
        version:
          type: integer
        audience_percent:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        targeting:
          $ref: "#/components/schemas/RuleNode"
        guardrails:
          type: array
          items:
            $ref: "#/components/schemas/GuardrailRule"
        versions:
          type: array
          items:
            type: object
        review_history:
          type: array
          items:
            type: object
        guardrail_log:
          type: array
          items:
            type: object
        decision_comment:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DecideRequest:
      type: object
      properties:
        subject_id:
          type: string
        attributes:
          type: object
          additionalProperties: true
        flag_keys:
          type: array
          items:
            type: string
    Decision:
      type: object
      properties:
        decision_id:
          type: string
        subject_id:
          type: string
        flag_key:
          type: string
        value:
          description: JSON literal
        experiment_id:
          type: string
        variant_name:
          type: string
        config_version:
          type: integer
        created_at:
          type: string
          format: date-time
    EventIngest:
      type: object
      required: [event_id, type, subject_id]
      properties:
        event_id:
          type: string
        type:
          type: string
        subject_id:
          type: string
        decision_id:
          type: string
        occurred_at:
          type: string
          format: date-time
          description: Если не задано — время приёма на сервере
        properties:
          type: object
          additionalProperties: true
    EventBatchResult:
      type: object
      properties:
        accepted:
          type: integer
        duplicate:
          type: integer
        rejected:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    ApproverPolicy:
      type: object
      properties:
        experimenter_id:
          type: string
        approver_ids:
          type: array
          items:
            type: string
        min_approvals:
          type: integer
    EventType:
      type: object
      required: [key]
      properties:
        key:
          type: string
        description:
          type: string
        requires_exposure:
          type: boolean
        archived:
          type: boolean
...